+++ CHANGELOG.md
@@ -13,6 +13,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `PlatformDependentCastRule` analysis rule, which flags casts between platform-dependent and
   platform-independent types.
 
+### Changed
+
+- Improve handling of cast expressions in `MemoryManagementRule`.
+- Change `memoryFunctions` parameter to fully-qualified names in `MemoryManagementRule`.
+- Remove `MemoryManagementRule` from the Sonar Way quality profile.
+
 ## [0.28.0] - 2021-02-02
 
 ### Added